CORE-19861 - attempt fix failure counter

+++ libs/state-manager/state-manager-db-impl/src/main/kotlin/net/corda/libs/statemanager/impl/StateManagerImpl.kt
@@ -85,7 +85,7 @@ class StateManagerImpl(
             )
         }
 
-        metricsRecorder.recordBatchSize(CREATE, states.size.toDouble())
+        metricsRecorder.recordBatchSize(CREATE, states.size)
 
         return metricsRecorder.recordProcessingTime(CREATE) {
             val successfulKeys = dataSource.connection.transaction { connection ->
@@ -111,7 +111,7 @@ class StateManagerImpl(
     override fun update(states: Collection<State>): Map<String, State?> {
         if (states.isEmpty()) return emptyMap()
 
-        metricsRecorder.recordBatchSize(UPDATE, states.size.toDouble())
+        metricsRecorder.recordBatchSize(UPDATE, states.size)
 
         return metricsRecorder.recordProcessingTime(UPDATE) {
             try {
@@ -138,7 +138,7 @@ class StateManagerImpl(
     override fun delete(states: Collection<State>): Map<String, State> {
         if (states.isEmpty()) return emptyMap()
 
-        metricsRecorder.recordBatchSize(DELETE, states.size.toDouble())
+        metricsRecorder.recordBatchSize(DELETE, states.size)
 
         return metricsRecorder.recordProcessingTime(DELETE) {
             try {
