+++ libs/state-manager/state-manager-db-impl/src/main/kotlin/net/corda/libs/statemanager/impl/metrics/MetricsRecorder.kt
@@ -13,7 +13,7 @@ interface MetricsRecorder {
 
     fun <T> recordProcessingTime(operationType: OperationType, block: () -> T): T
 
-    fun recordBatchSize(operationType: OperationType, size: Double)
+    fun recordBatchSize(operationType: OperationType, size: Int)
 
     fun recordFailureCount(operationType: OperationType, count: Int)
 }
